+++ Front_End/Dashboards Front/Final/requirements.txt
@@ -1,0 +1,4 @@
+streamlit==0.89.0
+pandas==1.3.3
+psycopg2==2.9.2
+streamlit_shadcn_ui==0.0.7
